Add inter trial interval and output interval support
I added a clock to the experment.py script.
Once we call run_experiment() this clock gets activated/reset
such that it tracks the time since the beginning of the experiment.
The clock has two use cases:

1. It keeps track of our inter trial interval for which I
added a new function inter_trial_interval() which gets
called after feedback has been given. Right now the ITI constant
is set to 200ms, so the ITI is 200ms long.
It has been ensured that the window is
getting flipped such that the participant cannot see
the feedback anymore. Moreover, key handling has been
ensured during this ITI which means we can close the window
in that time as well.

2. It keeps track of the output interval.
Every time 10ms pass in the task-phase (see docs)
data should be collected - this shall not be implemented
before the output stream is merged (I delayed it a bit).
I already added the last_trial_time update to the
run_trial() function (line 490). This attribute
is reset every time a task-phase starts. We could
save the cumulated time and calculate the delta time
afterwards just like Van Dam did it in the Handeln course.