Add Arduino experiment integration
1. I added a new function vibrotactile_cue()
which reads out current trial data and uses
the VibrationController class to play the cue.
A message is displayed during this.
Importantly, this is another waiting function
and will not end until the cue is over.
During this and once more key handling has been ensured.

2. For the Arduino board I added a new parameter and
a constant for the cue time (right now at 200ms as discussed).
I expanded the main class by a comment which holds
the implementation to run the experiment with the
Arduino board. The Arduino board therefore must be
manually given to the experiment. The experiment can
also start without an Arduino board but every time a cue
gets played a RuntimeWarning will be issued that cue was suppressed.
When the trial phase is over (in run_experiment()) the Arduino
board is closed if available.
This implementation ensures awareness of the Arduino resources.