Refactor feedback and inter-trial interval
Feedback and inter-trial interval [ITI] have
been moved into the run_trial() function.
This is because the trial structure / one trial
flow that we are going to use in our experiment
includes those. Therefore, semantically it
makes sense to move those there. With it
the docs changed because now we have everything
trial related located in the run_trial function.

The run experiment function is now concerned
with phase states and explanations. This is a
better division of concers and aligns with our
experimental structure.