Add trial tutorial
I added 10 trials with an introduction to those in the beginning
of the experiment where participants can get to know the general
trial flow.

This has no debug mode as it would require a lot of changes. Either
we need to add a lot of parameters to already exisiting functions
or create new ones then. Since we do not collect any data and
just the work flow like the cues need to work I decided to leave
that out.

Either way I changed some existing functions:
    - the vibrotactile_cue function has a new optional parameter where we can decide the intensity
      for a cue played ourselves
    - trial_confirmation can be used before init trial with a new conditional