Refactor text prompts
Further abstraction to add texts that need to be confirmed.
Changed give_explanation to take account for different situations
that come up in our experiments.

This is an intermediate change and will be overwritten later
since we have another implementation of blitting images
to the screen to give different explanations.